Upgrade solution to .NET Core 2.2
