+++ bin/rhisis-cli.bat
@@ -1,3 +1,3 @@
 @echo off
-dotnet tools/Rhisis.CLI/netcoreapp2.0/Rhisis.CLI.dll %*
+dotnet tools/Rhisis.CLI/netcoreapp2.2/Rhisis.CLI.dll %*
 exit
