+++ src/Rhisis.CLI/Application.cs
@@ -1,10 +1,10 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
-using Rhisis.CLI.Commands;
+using Rhisis.CLI.Commands.Database;
 
 namespace Rhisis.CLI
 {
     [Command(ThrowOnUnexpectedArgument = false, Description = Program.Description)]
-    [Subcommand("database", typeof(DatabaseCommand))]
+    [Subcommand(typeof(DatabaseCommand))]
     public class Application
     {
         public void OnExecute(CommandLineApplication app, IConsole console)
